Move dedupue implementation to doctype manager and update pieces accordingly
