+++ lib/modules/apostrophe-docs/lib/api.js
@@ -3,6 +3,7 @@ var _ = require('@sailshq/lodash');
 var Promise = require('bluebird');
 
 module.exports = function(self, options) {
+
   self.enableCollection = function(callback) {
     self.apos.db.collection('aposDocs', function(err, collection) {
       self.db = collection;
@@ -466,84 +467,6 @@ module.exports = function(self, options) {
 
   self.emptyTrash = function(req, idOrCriteria, callback) {
     return self.deleteFromTrash(req, idOrCriteria, callback);
-  };
-
-  // After moving document to the trash, prefix any fields that have
-  // unique indexes so that other pieces are allowed to reuse
-  // those usernames, email addresses, etc.
-
-  self.deduplicateTrash = function(req, doc, callback) {
-    var prefix = 'deduplicate-' + doc._id + '-';
-    var $set = {};
-    _.each(self.getManager(doc.type).trashPrefixFields, function(name) {
-      if ((typeof doc[name]) !== 'string') {
-        // Presumably a sparse index
-        return;
-      }
-      if (doc[name].substr(0, prefix.length) !== prefix) {
-        $set[name] = prefix + doc[name];
-      }
-      // So methods called later, or extending this method, see the change in piece
-      doc[name] = $set[name];
-    });
-    if (_.isEmpty($set)) {
-      return setImmediate(callback);
-    }
-    return self.apos.docs.db.update({ _id: doc._id }, { $set: $set }, callback);
-  };
-
-  // When rescuing docs from the trash, attempt to un-prefix any fields
-  // that have unique indexes. However, if we then find it's been taken
-  // in the meanwhile, leave the prefix in place and leave it up to
-  // the user to resolve the issue.
-
-  self.deduplicateRescue = function(req, doc, callback) {
-    var prefix = 'deduplicate-' + doc._id + '-';
-    var $set = {};
-    _.each(self.getManager(doc.type).trashPrefixFields, function(name) {
-      if ((typeof doc[name]) !== 'string') {
-        // Presumably a sparse index
-        return;
-      }
-      $set[name] = doc[name].replace(prefix, '');
-    });
-    return async.series([
-      checkDuplicates,
-      update
-    ], callback);
-    function checkDuplicates(callback) {
-      return async.eachSeries(self.trashPrefixFields, checkOne, callback);
-    }
-    function update(callback) {
-      var action = {
-        $set: $set
-      };
-      // So methods called later, or extending this method, see the change in docs
-      _.assign(doc, $set);
-      if (_.isEmpty($set)) {
-        // Nothing to do
-        return setImmediate(callback);
-      }
-      return self.apos.docs.db.update({ _id: doc._id }, action, callback);
-    }
-    function checkOne(name, callback) {
-      var query = { type: self.name, _id: { $ne: doc._id } };
-      query[name] = $set[name];
-      if ($set[name] === '') {
-        // Assume sparse index if empty strings are seen; don't
-        // generate lots of weird prefix-only email addresses
-        return setImmediate(callback);
-      }
-      return self.apos.docs.db.findOne(query, { _id: 1 }, function(err, found) {
-        if (err) {
-          return callback(err);
-        }
-        if (found) {
-          delete $set[name];
-        }
-        return callback(null);
-      });
-    }
   };
 
   // Recursively visit every property of a doc,
