+++ lib/modules/apostrophe-pieces/lib/api.js
@@ -3,7 +3,6 @@ var _ = require('@sailshq/lodash');
 var Promise = require('bluebird');
 
 module.exports = function(self, options) {
-
   self.finalizeControls = function() {
     self.createControls = self.options.createControls || [
       {
@@ -395,7 +394,7 @@ module.exports = function(self, options) {
         if (err) {
           return callback(err);
         }
-        return self.apos.docs.deduplicateTrash(req, piece, callback);
+        return self.apos.docs.getManager(self.name).deduplicateTrash(req, piece, callback);
       });
     }
   };
@@ -423,7 +422,7 @@ module.exports = function(self, options) {
         if (err) {
           return callback(err);
         }
-        return self.apos.docs.deduplicateRescue(req, piece, callback);
+        return self.apos.docs.getManager(self.name).deduplicateRescue(req, piece, callback);
       });
     }
   };
